packed-backend: fsck should warn when "packed-refs" file is empty
During fsck, an empty "packed-refs" gives an error; this is unwarranted.
The runtime code paths would accept an empty "packed-refs" file, such as
"create_snapshot" would simply return the "snapshot" without checking
the content of "packed-refs".

It might seem that we should skip checking if "packed-refs" file is
empty during fsck, which aligns with the runtime behavior. However, we
should think about whether we need to report a warning to the use when
"packed-refs" file is empty.

After 694b7a1999 (repack_without_ref(): write peeled refs in the
rewritten file, 2013-04-22), we would always write header into the
"packed-refs" file. And we would never create empty file since then. As
we may tighten this rule in the future, we should try to report this to
the users.

And we also need to decide the fsck message type, we should not choose
"FSCK_ERROR" due to we create such empty file in the very early
versions. And we should not break the compatibility. So, let's choose
"FSCK_INFO" as the fsck message type.

Create a new fsck message id "EMPTY_PACKED_REFS_FILE(INFO)" to indicate
that "packed-refs" is empty. Then, update the test to exercise the code.

Signed-off-by: shejialuo <shejialuo@gmail.com>