packed-backend: skip checking consistency of empty packed-refs file
In "load_contents", when the "packed-refs" is empty, we will just return
the snapshot. However, we would report an error to the user when
checking the consistency of the empty "packed-refs".

We should align with the runtime behavior. As what "load_contents" does,
let's check whether the file size is zero and if so, we will skip
checking the consistency and simply return.

Signed-off-by: shejialuo <shejialuo@gmail.com>